Visual tests now work as expected
