+++ libs/components/src/lib/button/ui.test.ts
@@ -1,54 +1,10 @@
 import * as path from 'path';
 import { expect, Page, test } from '@playwright/test';
-import { extractHTMLBlocksFromReadme } from '../../../scripts/visual-tests/utils';
-
-const defaultStyles = [
-	'http://127.0.0.1:8080/dist/core/fonts/spezia.css',
-	'http://127.0.0.1:8080/dist/core/theme/light.css'
-];
-
-/**
- * @param root0
- * @param root0.page
- * @param root0.componentName
- * @param root0.styleUrls
- */
-async function loadComponent({
-	page,
-	componentName,
-	styleUrls = defaultStyles,
-}: {
-	page: Page,
-	componentName: string,
-	styleUrls?: string[]
-}) {
-	await page.goto('http://127.0.0.1:8080/scripts/visual-tests/index.html');
-
-	await page.addScriptTag({
-		url: `http://127.0.0.1:8080/dist/components/${componentName}/index.js`,
-		type: 'module',
-	});
-
-	const styleTags$ = styleUrls.map(url => page.addStyleTag({ url }));
-	await Promise.all(styleTags$);
-}
-
-/**
- * @param root0
- * @param root0.page
- * @param root0.template
- */
-async function loadTemplate({
-	page,
-	template,
-}: { page: Page, template: string }) {
-	const wrappedTemplate = `<div id="wrapper">${template}</div>`;
-	await page.addScriptTag({
-		content: `
-            document.body.innerHTML = \`${wrappedTemplate}\`;
-        `,
-	});
-}
+import {
+  extractHTMLBlocksFromReadme,
+  loadComponent,
+  loadTemplate
+} from '../common/test-utils/visual-tests-utils';
 
 const componentName = 'button';
 test('should have all connotations', async ({ page }: { page: Page }) => {
