+++ package.json
@@ -15,10 +15,8 @@
     "@microsoft/fast-element": "^1.6.2",
     "@microsoft/fast-foundation": "^2.29.0",
     "@microsoft/fast-web-utilities": "^5.0.2",
-    "@playwright/test": "^1.16.3",
     "core-js": "^3.6.5",
     "eleventy-plugin-code-block-demo": "^0.1.6",
-    "playwright": "^1.16.3",
     "ramda": "^0.27.1",
     "regenerator-runtime": "0.13.7",
     "tslib": "^2.0.0"
@@ -36,6 +34,7 @@
     "@nrwl/tao": "13.4.1",
     "@nrwl/web": "13.4.1",
     "@nrwl/workspace": "13.4.1",
+    "@playwright/test": "1.16.3",
     "@rollup/plugin-node-resolve": "^13.1.3",
     "@rollup/plugin-typescript": "^8.3.0",
     "@types/jasmine": "~3.5.0",
@@ -49,6 +48,8 @@
     "eslint": "8.2.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
+    "extract-gfm": "^0.1.0",
+    "http-server": "^14.1.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
     "jest": "27.2.3",
@@ -58,6 +59,7 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
+    "playwright": "1.16.3",
     "prettier": "^2.3.1",
     "rollup-plugin-delete": "^2.0.0",
     "rollup-plugin-lit-css": "^3.2.1",
